[Improvement][api] Introduce taskrecord, udffunc, workergroup, lineage interface for clear code
