+++ dolphinscheduler-api/src/test/java/org/apache/dolphinscheduler/api/service/WorkFlowLineageServiceTest.java
@@ -16,11 +16,20 @@
  */
 package org.apache.dolphinscheduler.api.service;
 
+import static org.mockito.Mockito.when;
+
+import org.apache.dolphinscheduler.api.service.impl.WorkFlowLineageServiceImpl;
 import org.apache.dolphinscheduler.common.Constants;
-import org.apache.dolphinscheduler.common.utils.EncryptionUtils;
 import org.apache.dolphinscheduler.dao.entity.WorkFlowLineage;
 import org.apache.dolphinscheduler.dao.entity.WorkFlowRelation;
 import org.apache.dolphinscheduler.dao.mapper.WorkFlowLineageMapper;
+
+import java.util.ArrayList;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+
 import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -28,29 +37,27 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 
-import java.util.*;
-
-import static org.mockito.Mockito.when;
-
 @RunWith(MockitoJUnitRunner.class)
 public class WorkFlowLineageServiceTest {
 
     @InjectMocks
-    private WorkFlowLineageService workFlowLineageService;
+    private WorkFlowLineageServiceImpl workFlowLineageService;
 
     @Mock
     private WorkFlowLineageMapper workFlowLineageMapper;
 
     @Test
+    @SuppressWarnings("unchecked")
     public void testQueryWorkFlowLineageByName() {
         String searchVal = "test";
         when(workFlowLineageMapper.queryByName(searchVal, 1)).thenReturn(getWorkFlowLineages());
-        Map<String, Object> result = workFlowLineageService.queryWorkFlowLineageByName(searchVal,1);
-        List<WorkFlowLineage> workFlowLineageList = (List<WorkFlowLineage>)result.get(Constants.DATA_LIST);
-        Assert.assertTrue(workFlowLineageList.size()>0);
+        Map<String, Object> result = workFlowLineageService.queryWorkFlowLineageByName(searchVal, 1);
+        List<WorkFlowLineage> workFlowLineageList = (List<WorkFlowLineage>) result.get(Constants.DATA_LIST);
+        Assert.assertTrue(workFlowLineageList.size() > 0);
     }
 
     @Test
+    @SuppressWarnings("unchecked")
     public void testQueryWorkFlowLineageByIds() {
 
         Set<Integer> ids = new HashSet<>();
@@ -59,12 +66,12 @@ public class WorkFlowLineageServiceTest {
 
         when(workFlowLineageMapper.queryByIds(ids, 1)).thenReturn(getWorkFlowLineages());
         when(workFlowLineageMapper.querySourceTarget(1)).thenReturn(getWorkFlowRelation());
-        Map<String, Object> result = workFlowLineageService.queryWorkFlowLineageByIds(ids,1);
-        Map<String, Object> workFlowLists = (Map<String, Object>)result.get(Constants.DATA_LIST);
-        List<WorkFlowLineage> workFlowLineages = (List<WorkFlowLineage>)workFlowLists.get("workFlowList");
-        List<WorkFlowRelation> workFlowRelations = (List<WorkFlowRelation>)workFlowLists.get("workFlowRelationList");
-        Assert.assertTrue(workFlowLineages.size()>0);
-        Assert.assertTrue(workFlowRelations.size()>0);
+        Map<String, Object> result = workFlowLineageService.queryWorkFlowLineageByIds(ids, 1);
+        Map<String, Object> workFlowLists = (Map<String, Object>) result.get(Constants.DATA_LIST);
+        List<WorkFlowLineage> workFlowLineages = (List<WorkFlowLineage>) workFlowLists.get("workFlowList");
+        List<WorkFlowRelation> workFlowRelations = (List<WorkFlowRelation>) workFlowLists.get("workFlowRelationList");
+        Assert.assertTrue(workFlowLineages.size() > 0);
+        Assert.assertTrue(workFlowRelations.size() > 0);
     }
 
     private List<WorkFlowLineage> getWorkFlowLineages() {
@@ -76,7 +83,7 @@ public class WorkFlowLineageServiceTest {
         return workFlowLineages;
     }
 
-    private List<WorkFlowRelation> getWorkFlowRelation(){
+    private List<WorkFlowRelation> getWorkFlowRelation() {
         List<WorkFlowRelation> workFlowRelations = new ArrayList<>();
         WorkFlowRelation workFlowRelation = new WorkFlowRelation();
         workFlowRelation.setSourceWorkFlowId(1);
