+++ dolphinscheduler-api/src/test/java/org/apache/dolphinscheduler/api/service/WorkerGroupServiceTest.java
@@ -16,18 +16,21 @@
  */
 package org.apache.dolphinscheduler.api.service;
 
-import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
-import org.apache.dolphinscheduler.api.enums.Status;
+import org.apache.dolphinscheduler.api.service.impl.WorkerGroupServiceImpl;
 import org.apache.dolphinscheduler.api.utils.PageInfo;
 import org.apache.dolphinscheduler.common.Constants;
 import org.apache.dolphinscheduler.common.enums.UserType;
-import org.apache.dolphinscheduler.common.utils.CollectionUtils;
 import org.apache.dolphinscheduler.dao.entity.ProcessInstance;
 import org.apache.dolphinscheduler.dao.entity.User;
 import org.apache.dolphinscheduler.dao.entity.WorkerGroup;
-import org.apache.dolphinscheduler.dao.mapper.ProcessInstanceMapper;
 import org.apache.dolphinscheduler.service.zk.ZookeeperCachedOperator;
 import org.apache.dolphinscheduler.service.zk.ZookeeperConfig;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
@@ -35,38 +38,24 @@ import org.junit.runner.RunWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.Mockito;
-import org.mockito.internal.matchers.Any;
 import org.mockito.junit.MockitoJUnitRunner;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
 
 @RunWith(MockitoJUnitRunner.class)
 public class WorkerGroupServiceTest {
 
-    private static final Logger logger = LoggerFactory.getLogger(WorkerGroupServiceTest.class);
-
     @InjectMocks
-    private WorkerGroupService workerGroupService;
-
-    @Mock
-    private ProcessInstanceMapper processInstanceMapper;
+    private WorkerGroupServiceImpl workerGroupService;
 
     @Mock
     private ZookeeperCachedOperator zookeeperCachedOperator;
 
-
     @Before
-    public void init(){
+    public void init() {
         ZookeeperConfig zookeeperConfig = new ZookeeperConfig();
         zookeeperConfig.setDsRoot("/dolphinscheduler_qzw");
         Mockito.when(zookeeperCachedOperator.getZookeeperConfig()).thenReturn(zookeeperConfig);
 
-        String workerPath = zookeeperCachedOperator.getZookeeperConfig().getDsRoot()+"/nodes" +"/worker";
+        String workerPath = zookeeperCachedOperator.getZookeeperConfig().getDsRoot() + "/nodes" + "/worker";
 
         List<String> workerGroupStrList = new ArrayList<>();
         workerGroupStrList.add("default");
@@ -83,20 +72,22 @@ public class WorkerGroupServiceTest {
     }
 
     /**
-     *  query worker group paging
+     * query worker group paging
      */
     @Test
-    public void testQueryAllGroupPaging(){
+    @SuppressWarnings("unchecked")
+    public void testQueryAllGroupPaging() {
         User user = new User();
         // general user add
         user.setUserType(UserType.ADMIN_USER);
         Map<String, Object> result = workerGroupService.queryAllGroupPaging(user, 1, 10, null);
-        PageInfo<WorkerGroup> pageInfo = (PageInfo) result.get(Constants.DATA_LIST);
-        Assert.assertEquals(pageInfo.getLists().size(),1);
+        PageInfo<WorkerGroup> pageInfo = (PageInfo<WorkerGroup>) result.get(Constants.DATA_LIST);
+        Assert.assertEquals(pageInfo.getLists().size(), 1);
     }
 
 
     @Test
+    @SuppressWarnings("unchecked")
     public void testQueryAllGroup() throws Exception {
         Map<String, Object> result = workerGroupService.queryAllGroup();
         Set<String> workerGroups = (Set<String>) result.get(Constants.DATA_LIST);
@@ -106,9 +97,10 @@ public class WorkerGroupServiceTest {
 
     /**
      * get processInstances
-     * @return
+     *
+     * @return mock process instance list
      */
-    private List<ProcessInstance> getProcessInstanceList(){
+    private List<ProcessInstance> getProcessInstanceList() {
 
         List<ProcessInstance> processInstances = new ArrayList<>();
         processInstances.add(new ProcessInstance());
